LayoutTests/imported/w3c/web-platform-tests/user-timing/test_user_timing_mark_and_measure_exception_when_invoke_with_timing_attributes-expected.txt

Description

This test validates exception scenarios of invoking mark() and measure() with timing attributes as value.


PASS window.performance is defined 
FAIL performance.mark and performance.measure should throw if used with timing attribute values assert_throws: function "function () { window.performance.measure(timingAttributes..." did not throw
FAIL performance.mark and performance.measure should not throw if used with timing attribute values in workers Can't find variable: performance
